feat(gguf): guess default config file from files
